move *Expression classes into the expression package with the other *Expression classes
