+++ api/src/main/java/jakarta/data/metamodel/BasicAttribute.java
@@ -17,6 +17,7 @@
  */
 package jakarta.data.metamodel;
 
+import jakarta.data.metamodel.expression.Expression;
 import jakarta.data.metamodel.impl.BasicAttributeRecord;
 
 import java.util.Objects;
