+++ api/src/main/java/jakarta/data/metamodel/ComparableAttribute.java
@@ -17,6 +17,7 @@
  */
 package jakarta.data.metamodel;
 
+import jakarta.data.metamodel.expression.ComparableExpression;
 import jakarta.data.metamodel.impl.ComparableAttributeRecord;
 
 import java.util.Objects;
