+++ api/src/main/java/jakarta/data/metamodel/expression/ComparableExpression.java
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.metamodel;
+package jakarta.data.metamodel.expression;
 
 import jakarta.data.metamodel.constraint.Between;
 import jakarta.data.metamodel.constraint.GreaterThan;
