+++ api/src/main/java/jakarta/data/metamodel/expression/Expression.java
@@ -15,7 +15,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package jakarta.data.metamodel;
+package jakarta.data.metamodel.expression;
 
 import jakarta.data.metamodel.constraint.Constraint;
 import jakarta.data.metamodel.constraint.In;
