+++ api/src/main/java/jakarta/data/metamodel/expression/FunctionExpression.java
@@ -17,8 +17,6 @@
  */
 package jakarta.data.metamodel.expression;
 
-import jakarta.data.metamodel.Expression;
-
 import java.util.List;
 
 public interface FunctionExpression<T,V> extends Expression<T,V> {
