+++ api/src/main/java/jakarta/data/metamodel/expression/NumericOperatorExpression.java
@@ -17,9 +17,6 @@
  */
 package jakarta.data.metamodel.expression;
 
-import jakarta.data.metamodel.Expression;
-import jakarta.data.metamodel.NumericExpression;
-
 public interface NumericOperatorExpression<T, N extends Number & Comparable<N>>
         extends NumericExpression<T,N> {
     enum Operator {
