+++ api/src/main/java/jakarta/data/metamodel/path/ComparablePathRecord.java
@@ -18,7 +18,7 @@
 package jakarta.data.metamodel.path;
 
 import jakarta.data.metamodel.ComparableAttribute;
-import jakarta.data.metamodel.NavigableExpression;
+import jakarta.data.metamodel.expression.NavigableExpression;
 
 record ComparablePathRecord<T,U,C extends Comparable<?>>
         (NavigableExpression<T,U> expression, ComparableAttribute<U, C> attribute)
