+++ api/src/main/java/jakarta/data/metamodel/path/TextPathRecord.java
@@ -17,8 +17,8 @@
  */
 package jakarta.data.metamodel.path;
 
-import jakarta.data.metamodel.NavigableExpression;
 import jakarta.data.metamodel.TextAttribute;
+import jakarta.data.metamodel.expression.NavigableExpression;
 
 record TextPathRecord<T,U>
         (NavigableExpression<T,U> expression, TextAttribute<U> attribute)
